Major refactoring of target configuration, SCHEMA
The SCHEMA has been updated with some extensions in the request section.

conf/webroot-path and conf/rsa-key-size are now deprecated in favour of
configuration settings in the request section of target files; use the
default target file.

The quickstart wizard has been updated appropriately. The old files will
continue to be read for compatibility, but the values in the default
target file will take precedence if found.