Bugfix: import-le command was broken
This was caused by the refactoring to the storage module which changed
ImportKey to take a crypto.PrivateKey. But since this is an interface{},
it in practice accepts anything, so type checking didn't catch that
the import-le call to ImportKey needed to be updated. This caused
acmetool to panic when import-le was used.

The import-le command has also been enhanced so that it sets the
default provider URL if one is not set and there is only one directory
URL used in the imported state directory.

Fixes #86