Fix bug causing webroot path in legacy config file not to be used
This bug was caused by loading the legacy configuration file after
the targets were loaded, but default configuration is copied over
to the targets when the targets are loaded, so the configured
webroots were never used.

A workaround is to move the webroot paths to the conf/target file,
the now-preferred location for them. This fixes this bug, making
the workaround no longer necessary.

Fixes #85.