Fix hook scripts on systems with dash as /bin/sh
Previously, the hook scripts used /bin/sh as their shell.  This was then
changed to /bin/bash when it was reported that the scripts broke when
using /bin/sh on some systems.

However, not all systems have bash. The shell was thus changed back to
/bin/sh. It was determined that the breakage when using /bin/sh on some
systems was caused by /bin/sh being symlinked to /bin/dash, a more
conservative shell which does not support some bashisms. Namely, the
bashism of allowing "==" in [ ] rather than "=" was present in the
default scripts. This has been corrected, and the hook scripts now run
correctly under dash.

Fixes #105.

©! I hereby licence these changes under the licence with SHA256 hash
©! fd80a26fbb3f644af1fa994134446702932968519797227e07a1368dea80f0bc.