Change HAProxy hook to generic combined certificate hook
The hook is now installed by default, but only does anything if it
detects a daemon which needs it in $PATH or it is explicitly configured
to operate. Now detects quasselcore.

Quickstart no longer prompts whether to install the combined file
generation hook unless --expert is specified. Installation is performed
by default. Answers can still be given in response files.

The hook, combined certificate and private key files and the combined
certificate symlink directory are still all named haproxy for
compatibility.

Fixes #107.

©! I hereby licence these changes under the licence with SHA256 hash
©! fd80a26fbb3f644af1fa994134446702932968519797227e07a1368dea80f0bc.