Fix FDB bugs.
Fixes a bug where newly created symlinks were not chowned. This led
to acmetool constantly complaining about symlink owners not matching
when UID/GID enforcement is used.

Fixes a bug where newly created directories were not chowned when
UID/GID enforcement is used.

Fixes a bug where a symlink created in the tmp directory could make
acmetool inoperative if acmetool crashes before it is moved to its
proper place due to enforcement of symlink targets falling inside
the state directory.

Added code to clean the tmp directory of any files which are found in
it on startup.

©! I hereby licence these changes under the licence with SHA256 hash
©! fd80a26fbb3f644af1fa994134446702932968519797227e07a1368dea80f0bc.