Fix DNS challenge if the token starts with a hyphen
There is a small probability that the token stored in $CH_TXT_VALUE
starts with a hyphen. In these cases grep interprets the token as a
flag, instead of as the pattern, and fails.

This commit uses -- to force grep to interpret $CH_TXT_VALUE as the
pattern, even if it starts with a hyphen.