Allow hidden files within the state directory
Fixes #237.
Fixes #153.

This allows files or directories which have a path element beneath
the ACME state directory beginning with '.' to be ignored by the
permissions checker.

However, these files may not be referenced by symlinks which are not
themselves such files or directories, as the target of the symlink is
not permission-checked.

For example, this is allowed:

  /var/lib/acme
    .foo/
      bar.txt
    .hidden -> .foo (symlink)

This is not allowed:

  /var/lib/acme
    .foo/
      bar.txt
    unhidden -> .foo/bar.txt (symlink)

©! I, Hugo Landau <hlandau@devever.net>, hereby licence these changes under the
©! licence with SHA256 hash
©! fd80a26fbb3f644af1fa994134446702932968519797227e07a1368dea80f0bc.